requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.1
idna==2.10
imgurpython==1.1.7
multidict==5.1.0
nekos.py==1.0.3
Pillow==8.2.0
python-dotenv==0.17.0
pytube==10.7.1
requests==2.25.1
riotwatcher==3.1.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
